requirements.txt
alembic==1.18.4
blinker==1.9.0
click==8.3.1
Flask==3.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==3.21.1
marshmallow-sqlalchemy==1.1.0
packaging==26.0
python-dotenv==1.0.1
SQLAlchemy==2.0.46
typing_extensions==4.15.0
Werkzeug==3.1.5
